use Preferences class instead of EEPROM.
Instead of using a handcrafted data struct mapped to an EEPROM object,
use the Preferences library to store the configuration.
This also allows for using String() instead of char[] for hostname and
URL which is easier and more flexible.
The old configuration is updated at the first usage of the new code,
later on only the new Preferences values are used.